requirements.txt
streamlit
faiss-cpu
langchain
openai
pypdf
tiktoken
numpy